requirements.txt
tensorflow==2.11.0
streamlit==1.28.0
numpy==1.23.5
pillow==9.5.0
matplotlib==3.7.1
opencv-python==4.8.0.74
pandas==2.0.3
